# src/main/resources/static/openapi.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
  /api/v1/person/{id}:
    get:
      tags:
        - API para gerenciamento de pessoas
      summary: Buscar pessoa por id
      description: Retorna uma pessoa com base no id informado
      operationId: getPerson
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          required: true
          schema:
            type: string
        - name: size
          in: query
          required: true
          schema:
            type: string
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PersonResponseDTO"
    put:
      tags:
        - API para gerenciamento de pessoas
      summary: Atualizar pessoa
      description: Atualiza uma pessoa com base no id informado
      operationId: updatePerson
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePersonRequestDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PersonResponseDTO"
    delete:
      tags:
        - API para gerenciamento de pessoas
      summary: Deletar pessoa
      description: Deleta uma pessoa com base no id informado
      operationId: deletePerson
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: No Content
          content:
            '*/*':
              schema:
                type: object
  /api/v1/person:
    post:
      tags:
        - API para gerenciamento de pessoas
      summary: Criar pessoa
      description: Registra uma pessoa no banco de dados juntamente com seu endereço
        e o score
      operationId: createPerson
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePersonRequestDTO"
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PersonResponseDTO"
components:
  schemas:
    UpdatePersonRequestDTO:
      type: object
      properties:
        name:
          type: string
        age:
          type: string
        zipCode:
          type: string
        number:
          type: string
        score:
          type: integer
          format: int32
        ddd:
          type: string
        phoneNumber:
          type: string
    PersonResponseDTO:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        age:
          type: string
        zipCode:
          type: string
        state:
          type: string
        city:
          type: string
        neighborhood:
          type: string
        street:
          type: string
        number:
          type: string
        phone:
          type: string
        score:
          type: integer
          format: int32
    CreatePersonRequestDTO:
      type: object
      properties:
        name:
          type: string
        age:
          type: string
        zipCode:
          type: string
        number:
          type: string
        score:
          type: integer
          format: int32
        ddd:
          type: string
        phoneNumber:
          type: string
      required:
        - age
        - ddd
        - name
        - number
        - phoneNumber
        - score
        - zipCode